+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added support for percent width/height and fluid layouts
 * Improved load order of elements to reduce reflow.
 * Changed addEvent function name to 'on'. 
 * Removed conflicting array.indexOf function